[APL-1238] Move IteratorToStreamConverter to core converter package
