+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/converter/IteratorToStreamConverter.java
@@ -2,7 +2,7 @@
  * Copyright © 2018-2020 Apollo Foundation
  */
 
-package com.apollocurrency.aplwallet.apl.core.alias.converter;
+package com.apollocurrency.aplwallet.apl.core.converter;
 
 import com.apollocurrency.aplwallet.apl.core.db.DbIterator;
 import com.apollocurrency.aplwallet.apl.core.rest.converter.Converter;
